% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxspline1.R
\name{boxspline1}
\alias{boxspline1}
\title{Calculate 1D box-spline}
\usage{
boxspline1(x, Y, df, s, order, lim = range(x))
}
\arguments{
\item{x}{values at which to evaluate the splines}

\item{Y}{(optional) the coeficients}

\item{df}{number of degrees of freedom}

\item{s}{the lengths of convolution intervals}

\item{order}{the order of the spline (equivalent to length(s))}

\item{lim}{the range in which the box-spline is constructed (default: range(x))}
}
\value{
If Y coeficients are provided then box-spline is returned. If not, a matrix of base spline functions is returned.
}
\description{
Calculates the 1D box-spline base functions, or interpolation.
}
\examples{
x = seq(0,1,len=100)
y = boxspline1(x)
matplot(x, y, type="l", lty=1, ylim=c(0,1))
y = boxspline1(x, df=10, order=2)
matplot(x, y, type="l", lty=1, ylim=c(0,1), col=rainbow(10))
y = boxspline1(x, df=10, s=c(0.1,0.3))
matplot(x, y, type="l", lty=1, ylim=c(0,1), col=rainbow(10))

Y = rnorm(10)
y = boxspline1(x, Y, order=0)
plot(x, y, type="l")
y = boxspline1(x, Y, order=1)
lines(x, y, type="l", col=2)
y = boxspline1(x, Y, order=2)
lines(x, y, type="l", col=3)
y = boxspline1(x, Y, order=3)
lines(x, y, type="l", col=4)

Y = rnorm(10)
y = boxspline1(x, Y, order=1)
plot(x, y, type="l",lty=2)
y = boxspline1(x, Y, s=c(0.1,0.1))
lines(x, y, type="l", col=1)
y = boxspline1(x, Y, s=c(0.2,0.1))
lines(x, y, type="l", col=2)
y = boxspline1(x, Y, s=c(0.3,0.1))
lines(x, y, type="l", col=3)
}

